+++ cspell.config.yaml
@@ -122,6 +122,7 @@ words:
   - Keyyable
   - kirakira
   - kwaa
+  - lemonneko
   - lemonnekogh
   - lerp
   - libsamplerate
