--- packages/stage-ui-three/src/bus/index.ts
@@ -1,3 +1,0 @@
-export * from './factory'
-export * from './sticky-bus'
-export * from './types'
