+++ packages/stage-ui-three/src/composables/index.ts
@@ -1,5 +1,2 @@
-export * from './animation'
-export * from './core'
-export * from './expression'
-export * from './loader'
-export * from './shader/ibl'
+export * from './shader'
+export * from './vrm'
