+++ packages/stage-ui-three/src/composables/shader/index.ts
@@ -1,0 +1,1 @@
+export * from './ibl'
