+++ packages/stage-ui-three/src/composables/vrm/index.ts
@@ -1,0 +1,4 @@
+export * from './animation'
+export * from './core'
+export * from './expression'
+export * from './loader'
