+++ packages/stage-ui-three/src/composables/vrm/loader.ts
