+++ packages/stage-ui-three/src/composables/vrm/utils/eye-motions.ts
